backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn==0.34.0

# Database
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10

# Validation & settings
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.1.1

# Auth & security
python-jose==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0


# HTTP client (LLM API calls)
httpx==0.28.1

# File uploads / forms
python-multipart==0.0.20

# Text-to-Speech
elevenlabs>=1.0.0
python-dotenv>=1.0.0
